Workaround shared_ptr usage in constructors.
It's not really possible to use shared_from_this in the constructor,
because the shared_ptr doesn't exist yet. We can create one in the
constructor, but that's unsafe because the caller would create a second
shared_ptr. Fortunately, in these cases, there are limited callers and
we can ensure they don't create an extra shared_ptr.